src/server/requirements.txt
Flask
Flask-Cors
bcrypt
PyJWT
python-dotenv
